Add policies, change index(), add show()

+++ app/Http/Controllers/ScheduleController.php
@@ -16,58 +16,24 @@ class ScheduleController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index(Request $request): Response
+    public function index(Request $request): Response | RedirectResponse
     {
-        $id = $request->get('id', null);
-        $sort = strtoupper($request->get('sort', 'DESC'));
-        $sort_by = strtolower($request->get('sort_by', 'schedule_detail_id'));
-        if ($sort_by == "dateofweek") $sort_by = "dateOfWeek";
-        if (!$id) {
-            $schedule = auth()->user()->schedules()->first();
-        } else {
-            if (auth()->user()->schedules()->where("schedule_id", $id)->exists()) {
-                $schedule = auth()->user()->schedules()->where("schedule_id", $id)->first();
-            } else {
-                $schedule = auth()->user()->schedules()->first();
-            }
+        if (auth()->user()->schedules()->count() > 0) {
+            return redirect(route('schedule.show', auth()->user()->schedules()->first()));
         }
-        if ($schedule) {
-            $details = $schedule->details();
-            $total_row = $details->count();
-            $limit = $request->get('limit', 20);
-            if ($limit <= 0) $limit = 20;
-            $page = $request->get('page', 1);
-            if ($page <= 0) $page = 1;
-            $total_page = ceil($total_row/$limit);
-            if ($page > $total_page) $page = $total_page;
-            return Inertia::render("Schedule", [
-                "schedules" => fn () => auth()->user()->schedules,
-                "schedule_selected" => fn () => $schedule,
-                "schedule_details" => fn () => $details->OrderBy($sort_by, $sort)->limit($limit)->skip(($page-1)*$limit)->get()->load(["subject"]),
-                "subjects" => fn () => auth()->user()->subjects,
-                "numOfClassPeriodsPerDay" => fn () => $schedule->numOfClassPeriodsPerDay,
-                'timeOfEachClassPeriod' => fn () => $schedule->timeOfEachClassPeriod,
-                "limit" => $limit,
-                "page" => $page,
-                "total_page" => $total_page,
-                "sort" => $sort,
-                "sort_by" => $sort_by,
-            ]);
-        } else {
-            return Inertia::render("Schedule", [
-                "schedules" => fn () => auth()->user()->schedules,
-                "schedule_selected" => fn () => null,
-                "schedule_details" => fn () => [],
-                "subjects" => fn () => auth()->user()->subjects,
-                "numOfClassPeriodsPerDay" => fn () => 0,
-                'timeOfEachClassPeriod' => fn () => [],
-                "limit" => 0,
-                "page" => 0,
-                "total_page" => 0,
-                "sort" => $sort,
-                "sort_by" => $sort_by
-            ]);
-        }
+        return Inertia::render("Schedule", [
+            "schedules" => fn () => [],
+            "schedule_selected" => fn () => null,
+            "schedule_details" => fn () => [],
+            "subjects" => fn () => auth()->user()->subjects,
+            "numOfClassPeriodsPerDay" => fn () => 0,
+            'timeOfEachClassPeriod' => fn () => [],
+            "limit" => 0,
+            "page" => 0,
+            "total_page" => 0,
+            "sort" => null,
+            "sort_by" => null
+        ]);
     }
 
     /**
@@ -95,9 +61,48 @@ class ScheduleController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(Schedule $schedule)
+    public function show(Request $request, Schedule $schedule): Response
     {
-        //
+        $this->authorize('view', $schedule);
+        // $id = $request->get('id', null);
+        $sort = strtoupper($request->get('sort', 'DESC'));
+        $sort_by = strtolower($request->get('sort_by', 'schedule_detail_id'));
+        if ($sort_by == "dateofweek") $sort_by = "dateOfWeek";
+        // if (!$id) {
+        //     $schedule = auth()->user()->schedules()->first();
+        // } else {
+        //     if (Schedule::find($id)) {
+
+        //     } else {
+        //         $this->authorize('viewAny');
+        //     }
+        //     if (auth()->user()->schedules()->where("schedule_id", $id)->exists()) {
+        //         $schedule = auth()->user()->schedules()->where("schedule_id", $id)->first();
+        //     } else {
+        //         $schedule = auth()->user()->schedules()->first();
+        //     }
+        // }
+        $details = $schedule->details();
+        $total_row = $details->count();
+        $limit = $request->get('limit', 20);
+        if ($limit <= 0) $limit = 20;
+        $page = $request->get('page', 1);
+        if ($page <= 0) $page = 1;
+        $total_page = ceil($total_row/$limit);
+        if ($page > $total_page) $page = $total_page;
+        return Inertia::render("Schedule", [
+            "schedules" => fn () => auth()->user()->schedules,
+            "schedule_selected" => fn () => $schedule,
+            "schedule_details" => fn () => $details->OrderBy($sort_by, $sort)->limit($limit)->skip(($page-1)*$limit)->get()->load(["subject"]),
+            "subjects" => fn () => auth()->user()->subjects,
+            "numOfClassPeriodsPerDay" => fn () => $schedule->numOfClassPeriodsPerDay,
+            'timeOfEachClassPeriod' => fn () => $schedule->timeOfEachClassPeriod,
+            "limit" => $limit,
+            "page" => $page,
+            "total_page" => $total_page,
+            "sort" => $sort,
+            "sort_by" => $sort_by,
+        ]);
     }
 
     /**
@@ -113,6 +118,8 @@ class ScheduleController extends Controller
      */
     public function update(UpdateRequest $request, Schedule $schedule): RedirectResponse
     {
+        $this->authorize('update', $schedule);
+
         $validated = $request->validated();
 
         $schedule->update($validated);
@@ -125,6 +132,8 @@ class ScheduleController extends Controller
      */
     public function destroy(Schedule $schedule): RedirectResponse
     {
+        $this->authorize('delete', $schedule);
+
         $schedule->delete();
 
         return redirect(route('schedule.index'));
